+++ src/App.vue
@@ -21,6 +21,7 @@
       <BoardManager v-if="appStore.showBoardManager" />
       <GitHubManager v-if="appStore.showGithubManager" />
       <UpdaterPanel v-if="appStore.showUpdater" />
+      <ImportUrlDialog v-if="appStore.showImportUrl" />
     </template>
   </v-app>
 </template>
@@ -44,6 +45,7 @@ import PluginManager from './components/PluginManager.vue'
 import BoardManager from './components/BoardManager.vue'
 import GitHubManager from './components/GitHubManager.vue'
 import UpdaterPanel from './components/UpdaterPanel.vue'
+import ImportUrlDialog from './components/ImportUrlDialog.vue'
 
 const appStore = useAppStore()
 const serialStore = useSerialStore()